feat:实现了A星寻路
1. 使用A星算法实现了寻路
2. 通过二叉树和栈结构优化了排序效率
3. 现存问题：拐角处通行容易卡住

+++ Assets/Scripts/EnemySetting.cs
@@ -5,9 +5,9 @@ using UnityEngine;
 public class EnemySetting : ScriptableObject
 {
     public float hearing = 0.03f;
-    public float vision = 5f;
-    public float fleeVision = 10f;
-    public float patrolVision = 10f;
+    public int vision = 5;
+    public int fleeVision = 10;
+    public int patrolVision = 10;
     public float maxHp = 10f;
     public float moveSpeed = 0.5f;
     public float angleSpeed = 10f;
